+++ test/src/test_mat4.h
@@ -48,7 +48,7 @@ TEST_IMPL(GLM_PREFIX, mat4_identity) {
 }
 
 TEST_IMPL(GLM_PREFIX, mat4_identity_array) {
-  int i, count;
+  size_t i, count;
   mat4 matrices[4] = {
     A_MATRIX,
     A_MATRIX,
